reworking the micm class to make it easier to add a cuda solver
