+++ include/musica/micm/parse.hpp
@@ -2,15 +2,14 @@
 
 #include <musica/micm/chemistry.hpp>
 #include <musica/micm/micm.hpp>
-#include <musica/util.hpp>
 
 #include <mechanism_configuration/parser.hpp>
 
 namespace musica
 {
-  Chemistry ReadConfiguration(const std::string& config_path, Error* error);
-  Chemistry ParserV0(const mechanism_configuration::ParserResult<>& result, Error* error);
-  Chemistry ParserV1(const mechanism_configuration::ParserResult<>& result, Error* error);
+  Chemistry ReadConfiguration(const std::string& config_path);
+  Chemistry ParserV0(const mechanism_configuration::ParserResult<>& result);
+  Chemistry ParserV1(const mechanism_configuration::ParserResult<>& result);
 
   // Utility functions to check types and perform conversions
   bool IsBool(const std::string& value);
