+++ src/micm/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources(musica
   PRIVATE
     micm.cpp
+    micm_c_interface.cpp
     state.cpp
     parse.cpp
     v0_parse.cpp
