+++ src/micm/parse.cpp
@@ -1,14 +1,10 @@
 #include <musica/micm/parse.hpp>
-#include <musica/util.hpp>
-
 #include <mechanism_configuration/parser.hpp>
 
 namespace musica
 {
-  Chemistry ReadConfiguration(const std::string& config_path, Error* error)
+  Chemistry ReadConfiguration(const std::string& config_path)
   {
-    DeleteError(error);
-
     mechanism_configuration::UniversalParser parser;
     Chemistry chemistry{};
 
@@ -20,7 +16,7 @@ namespace musica
       {
         errors += error.second + "\n";
       }
-      *error = ToError(MUSICA_ERROR_CATEGORY, MUSICA_ERROR_CODE_CONFIG_PARSE_FAILED, errors.c_str());
+      throw std::runtime_error(errors);
     }
     else
     {
@@ -28,11 +24,11 @@ namespace musica
 
       switch (version.major)
       {
-        case 0: chemistry = ParserV0(parsed, error); break;
-        case 1: chemistry = ParserV1(parsed, error); break;
+        case 0: chemistry = ParserV0(parsed); break;
+        case 1: chemistry = ParserV1(parsed); break;
         default:
           std::string msg = "Version " + std::to_string(version.major) + " not supported";
-          *error = ToError(MUSICA_ERROR_CATEGORY, MUSICA_ERROR_CODE_VERSION_NOT_SUPPORTED, msg.c_str());
+          throw std::runtime_error(msg);
       }
     }
